Completed Assignment
I completed this in-class assignment. I missed class, but did my best to
pick up on the topics and complete all of the assignment. My final
thresholds in Ilastik and MATLAB using hysteresis worked OK, but they
aren't perfect. I'm not sure how to improve them, but I plan on
continuing to tweak them.